chore: 🤖 ck point
