+++ crates/rolldown_stats/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "rolldown_stats"
+version = "0.1.0"
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+repository.workspace = true
+
+[dependencies]
+
+[lints]
+workspace = true
