+++ crates/rolldown_stats/src/lib.rs
@@ -1,0 +1,5 @@
+mod hook_metric;
+
+pub struct Stats {
+  pub hook_metric: Vec<hook_metric::HookMetric>,
+}
